Remove UserService from getDemo*() methods, add Fenix API keys constants, verify that authUser exists when checking course active users

+++ backend/src/main/java/pt/ulisboa/tecnico/socialsoftware/tutor/course/CourseService.java
@@ -130,6 +130,7 @@ public class CourseService {
     public void anonymizeCourseExecutionUsers(int executionId) {
         CourseExecution courseExecution = courseExecutionRepository.findById(executionId).orElseThrow(() -> new TutorException(COURSE_EXECUTION_NOT_FOUND));
         for (User user : courseExecution.getUsers()) {
+            if (user.getAuthUser() == null) continue;
             String oldUsername = user.getUsername();
             authUserRepository.delete(user.getAuthUser());
             user.setAuthUser(null);
